Review Account holder implementations
Existing instruments `AccountHolder` was renamed to `InstrumentAccountHolder` and extends
the base type define in common. Common type is not final anymore. Marketplace defined
incorrectly a `InstrumentAccountHolder` that was renamed to `MarketplaceAccountHolder.`
