+++ src/main/java/com/checkout/marketplace/MarketplaceAccountHolder.java
@@ -1,22 +1,18 @@
 package com.checkout.marketplace;
 
-import com.checkout.common.Address;
 import com.checkout.common.CountryCode;
+import com.checkout.common.four.AccountHolder;
 import com.google.gson.annotations.SerializedName;
-import lombok.Builder;
 import lombok.Data;
+import lombok.EqualsAndHashCode;
+import lombok.ToString;
 
 @Data
-@Builder
-public final class InstrumentAccountHolder {
+@EqualsAndHashCode(callSuper = true)
+@ToString(callSuper = true)
+public final class MarketplaceAccountHolder extends AccountHolder {
 
-    private InstrumentAccountHolderType type;
-
-    @SerializedName("first_name")
-    private String firstName;
-
-    @SerializedName("last_name")
-    private String lastName;
+    private MarketplaceAccountHolderType type;
 
     @SerializedName("company_name")
     private String companyName;
@@ -30,12 +26,8 @@ public final class InstrumentAccountHolder {
     @SerializedName("country_of_birth")
     private CountryCode countryOfBirth;
 
+    @SerializedName("residential_status")
     private String residentialStatus;
-
-    @SerializedName("billing_address")
-    private Address billingAddress;
-
-    private Phone phone;
 
     private Identification identification;
