+++ src/main/java/com/checkout/marketplace/MarketplaceAccountHolderType.java
@@ -2,7 +2,7 @@ package com.checkout.marketplace;
 
 import com.google.gson.annotations.SerializedName;
 
-public enum InstrumentAccountHolderType {
+public enum MarketplaceAccountHolderType {
 
     @SerializedName("individual")
     INDIVIDUAL,
